Remove unnecessary unboxing in DateRenderer

+++ src/org/stringtemplate/v4/DateRenderer.java
@@ -72,7 +72,7 @@ public class DateRenderer implements AttributeRenderer<Object> {
         DateFormat f;
         if ( styleI==null ) f = new SimpleDateFormat(formatString, locale);
         else {
-            int style = styleI.intValue();
+            int style = styleI;
             if ( formatString.startsWith("date:") ) f = DateFormat.getDateInstance(style, locale);
             else if ( formatString.startsWith("time:") ) f = DateFormat.getTimeInstance(style, locale);
             else f = DateFormat.getDateTimeInstance(style, style, locale);
